# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health/live": {
      "get": {
        "operationId": "HealthController_live",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Процесс жив",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Проверить, что процесс API работает",
        "tags": [
          "health"
        ]
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "HealthController_ready",
        "parameters": [],
        "responses": {
          "200": {
            "description": "API принимает трафик",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "API ещё не слушает, уже останавливается или база данных недоступна",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Проверить готовность API принимать трафик",
        "tags": [
          "health"
        ]
      }
    },
    "/demo/conductor/{conductorSecret}": {
      "get": {
        "operationId": "DemoConductorController_show",
        "parameters": [
          {
            "name": "conductorSecret",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConductorSnapshotResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Секрет пульта неизвестен",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Снимок живой заявки для пульта ведущего",
        "tags": [
          "demo"
        ]
      }
    },
    "/demo/conductor/{conductorSecret}/advance": {
      "post": {
        "operationId": "DemoConductorController_advance",
        "parameters": [
          {
            "name": "conductorSecret",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConductorSnapshotResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Секрет пульта неизвестен",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "Живая заявка уже на последнем шаге",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Один переход автомата живой заявки",
        "tags": [
          "demo"
        ]
      }
    },
    "/demo/conductor/{conductorSecret}/reset": {
      "post": {
        "operationId": "DemoConductorController_reset",
        "parameters": [
          {
            "name": "conductorSecret",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConductorSnapshotResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Секрет пульта неизвестен",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Сброс живой заявки к принятой",
        "tags": [
          "demo"
        ]
      }
    },
    "/demo/links": {
      "get": {
        "operationId": "DemoLinksController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoLinksResponseDto"
                }
              }
            }
          }
        },
        "summary": "Служебный список демо-ссылок на заявки",
        "tags": [
          "demo"
        ]
      }
    },
    "/requests/{accessSecret}": {
      "get": {
        "operationId": "RequestPortalController_show",
        "parameters": [
          {
            "name": "accessSecret",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestPortalResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Заявка не найдена по секрету",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Слишком много запросов с этого IP",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds until the request may be retried",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          }
        },
        "summary": "Кабинет одной заявки по секрету в URL",
        "tags": [
          "requests"
        ]
      }
    }
  },
  "info": {
    "title": "ПК «Нордщит» — кабинет заявки",
    "description": "",
    "version": "0.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "schemas": {
      "HealthDataDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok"
            ],
            "example": "ok"
          }
        },
        "required": [
          "status"
        ]
      },
      "TraceMetaDto": {
        "type": "object",
        "properties": {
          "traceId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "traceId"
        ]
      },
      "HealthResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/HealthDataDto"
          },
          "meta": {
            "$ref": "#/components/schemas/TraceMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "ConductorSnapshotDataDto": {
        "type": "object",
        "properties": {
          "publicNumber": {
            "type": "string",
            "example": "З-10046"
          },
          "status": {
            "type": "string",
            "enum": [
              "accepted",
              "in_calculation",
              "quote_ready",
              "invoice_issued"
            ],
            "example": "accepted"
          },
          "statusLabel": {
            "type": "string",
            "example": "Принят"
          },
          "portalPath": {
            "type": "string",
            "example": "/r/seed-z10046-live-severnaya-duga"
          },
          "nextStatus": {
            "type": "string",
            "enum": [
              "accepted",
              "in_calculation",
              "quote_ready",
              "invoice_issued"
            ],
            "example": "in_calculation",
            "nullable": true
          },
          "nextStatusLabel": {
            "type": "string",
            "example": "В расчёте",
            "nullable": true
          }
        },
        "required": [
          "publicNumber",
          "status",
          "statusLabel",
          "portalPath",
          "nextStatus",
          "nextStatusLabel"
        ]
      },
      "ConductorSnapshotResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ConductorSnapshotDataDto"
          },
          "meta": {
            "$ref": "#/components/schemas/TraceMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "DemoLinkItemDto": {
        "type": "object",
        "properties": {
          "publicNumber": {
            "type": "string",
            "example": "З-10041"
          },
          "counterpartyName": {
            "type": "string",
            "example": "ООО «Северэнергомонтаж»"
          },
          "title": {
            "type": "string",
            "example": "Щит ЩО-70 800 А"
          },
          "status": {
            "type": "string",
            "enum": [
              "accepted",
              "in_calculation",
              "quote_ready",
              "invoice_issued"
            ],
            "example": "accepted"
          },
          "statusLabel": {
            "type": "string",
            "example": "Принят"
          },
          "portalPath": {
            "type": "string",
            "example": "/r/seed-z10041-accepted-severenergo"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-09-01T10:00:00.000Z",
            "format": "date-time"
          }
        },
        "required": [
          "publicNumber",
          "counterpartyName",
          "title",
          "status",
          "statusLabel",
          "portalPath",
          "updatedAt"
        ]
      },
      "DemoLinksDataDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DemoLinkItemDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "DemoLinksResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DemoLinksDataDto"
          },
          "meta": {
            "$ref": "#/components/schemas/TraceMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "RequestStageDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "accepted",
              "in_calculation",
              "quote_ready",
              "invoice_issued"
            ]
          },
          "label": {
            "type": "string",
            "example": "Принят"
          },
          "reachedAt": {
            "type": "string",
            "example": "2026-09-01T09:00:00.000Z",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "status",
          "label",
          "reachedAt"
        ]
      },
      "RequestSpecLineDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "example": 1,
            "format": "int32"
          },
          "unit": {
            "type": "string",
            "example": "шт"
          },
          "comment": {
            "type": "string",
            "example": "IP54, навесное"
          }
        },
        "required": [
          "name",
          "quantity",
          "unit"
        ]
      },
      "RequestFileDto": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "example": "КП-З-10043.pdf"
          },
          "kind": {
            "type": "string",
            "enum": [
              "questionnaire",
              "quote",
              "invoice"
            ]
          },
          "byteSize": {
            "type": "integer",
            "example": 240000,
            "format": "int32"
          },
          "uploadedAt": {
            "type": "string",
            "example": "2026-09-04T12:00:00.000Z",
            "format": "date-time"
          }
        },
        "required": [
          "fileName",
          "kind",
          "byteSize",
          "uploadedAt"
        ]
      },
      "RequestPortalDataDto": {
        "type": "object",
        "properties": {
          "publicNumber": {
            "type": "string",
            "example": "З-10043"
          },
          "counterpartyName": {
            "type": "string",
            "example": "ИП Кузнецов П.А."
          },
          "title": {
            "type": "string",
            "example": "ВРУ 400 А"
          },
          "status": {
            "type": "string",
            "enum": [
              "accepted",
              "in_calculation",
              "quote_ready",
              "invoice_issued"
            ],
            "example": "quote_ready"
          },
          "statusLabel": {
            "type": "string",
            "example": "КП готово"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-09-04T12:00:00.000Z",
            "format": "date-time"
          },
          "plantName": {
            "type": "string",
            "example": "ПК «Нордщит»"
          },
          "stages": {
            "maxItems": 4,
            "minItems": 4,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RequestStageDto"
            }
          },
          "specLines": {
            "maxItems": 5,
            "minItems": 2,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RequestSpecLineDto"
            }
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RequestFileDto"
            }
          },
          "demoLive": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "publicNumber",
          "counterpartyName",
          "title",
          "status",
          "statusLabel",
          "updatedAt",
          "plantName",
          "stages",
          "specLines",
          "files"
        ]
      },
      "RequestPortalResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RequestPortalDataDto"
          },
          "meta": {
            "$ref": "#/components/schemas/TraceMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string",
            "example": "One or more fields are invalid"
          },
          "errors": {
            "type": "array",
            "example": [
              {
                "field": "pageSize",
                "message": "pageSize must not be greater than 48"
              }
            ],
            "items": {
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "field",
                "message"
              ],
              "type": "object"
            }
          },
          "instance": {
            "type": "string",
            "example": "/api/v1/resources"
          },
          "status": {
            "type": "number",
            "example": 400
          },
          "title": {
            "type": "string",
            "example": "Validation failed"
          },
          "traceId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "example": "https://demo.local/problems/validation-error"
          }
        },
        "required": [
          "detail",
          "instance",
          "status",
          "title",
          "traceId",
          "type"
        ]
      }
    }
  }
}